+++ src/components/Loader/Loader.tsx
@@ -1,0 +1,40 @@
+/**
+ * @author Sergey Kozlov <skozlov@usetech.com>
+ */
+
+import React, { ReactNode } from 'react';
+import classNames from 'classnames';
+import { DimentionType, PlacementType } from '../../types';
+import './Loader.scss';
+
+export interface LoaderProps {
+    isFullPage?: boolean;
+    label?: string | ReactNode | null;
+    size?: DimentionType;
+    placement?: PlacementType;
+}
+
+const Loader = ({
+    isFullPage = false,
+    placement = 'right',
+    size = 'small',
+    label
+}: LoaderProps) => (
+    <div
+        className={classNames('unique-loader', {
+            full: isFullPage,
+        })}
+        role={'progressbar'}
+    >
+        <div className={classNames('loader-content', placement)}>
+            <div className={classNames('loader', size)} />
+            {label && (
+                <div data-testid={'loader-label'} className={'loader-label'}>
+                    {label}
+                </div>
+            )}
+        </div>
+    </div>
+);
+
+export default Loader;