+++ src/components/Suggest/components/SuggestSpinner/SuggestSpinner.tsx
@@ -1,4 +1,5 @@
 import { ReactNode } from 'react';
+import Loader from '../../../Loader/Loader';
 
 import './SuggestSpinner.scss';
 
@@ -7,7 +8,7 @@ type SuggestSpinnerProps = {
 };
 
 export const SuggestSpinner = ({ loadingText }: SuggestSpinnerProps) => (
-    <div className={'suggest-spinner'} role={'progressbar'}>
-        {loadingText}
+    <div className={'suggest-spinner'}>
+        <Loader label={loadingText} size={'small'} />
     </div>
 );